CLN: introduced GrantaConnectionMixin, refactored base classes
