+++ granta_example/core/base_classes.py
@@ -1,12 +1,13 @@
-import granta
-
-from traits.trait_types import Any, Str
-
 from force_bdss.api import (
-    BaseDataSourceModel, BaseNotificationListener,
-    BaseNotificationListenerModel)
+    BaseDataSource, BaseDataSourceModel,
+    BaseNotificationListener, BaseNotificationListenerModel)
 
-from .granta_auth_mixin import GrantaAuthMixin
+from .granta_mixin_classes import GrantaAuthMixin, GrantaConnectMixin
+
+
+class BaseGrantaDataSource(GrantaConnectMixin, BaseDataSource):
+    """A BaseDataSource with utility traits and methods for
+    connecting to a Granta MI database"""
 
 
 class BaseGrantaDataSourceModel(GrantaAuthMixin, BaseDataSourceModel):
@@ -14,22 +15,13 @@ class BaseGrantaDataSourceModel(GrantaAuthMixin, BaseDataSourceModel):
     Granta MI authentication details"""
 
 
-class BaseGrantaNotificationListener(BaseNotificationListener):
-
-    #: Reference to the GrantaMI session
-    _mi = Any()
-
-    _db_key = Str()
+class BaseGrantaNotificationListener(
+        GrantaConnectMixin, BaseNotificationListener):
+    """A BaseNotificationListener with utility traits and methods for
+    connecting to a Granta MI database"""
 
     def initialize(self, model):
-        """Expects model to be of type
-        BaseGrantaNotificationListenerModel"""
-        self._mi = granta.MI(
-            model.url,
-            user_name=model.login,
-            password=model.password,
-            domain=model.domain)
-        self._db_key = model.db_key
+        self._connect_mi(model)
 
 
 class BaseGrantaNotificationListenerModel(
